Revert "Merge pull request #5725 from adrienthebo/revert-pup-7283"
This reverts commit 5671a3b21b67345fc8afcd61080bcd05f9e4924b, reversing
changes made to cc9465c31463a5f3a42f7198978f88a10087228a.
