+++ lib/puppet/indirector/certificate_revocation_list/rest.rb
@@ -7,16 +7,4 @@ class Puppet::SSL::CertificateRevocationList::Rest < Puppet::Indirector::REST
   use_server_setting(:ca_server)
   use_port_setting(:ca_port)
   use_srv_service(:ca)
-
-  def find(request)
-    if !Puppet::FileSystem.exist?(Puppet[:hostcrl])
-      msg =  "Disable certificate revocation checking when fetching the CRL and no CRL is present"
-      overrides = {certificate_revocation: false}
-      Puppet.override(overrides, msg) do
-        super
-      end
-    else
-      super
-    end
-  end
 end
